migrate test_servicenow to pytest
